refactor(ui): TMS-2508 tenant create rename
Moved the tenant create dialog into the tenant directory and renamed.
