+++ frontend/src/views/TenantListView.vue
@@ -3,7 +3,7 @@ import { storeToRefs } from 'pinia'
 import { onMounted, ref } from 'vue'
 import { useRouter } from 'vue-router'
 
-import CreateTenantDialog from '@/components/CreateTenantDialog.vue'
+import TenantCreateDialog from '@/components/tenant/TenantCreateDialog.vue'
 import TenantList from '@/components/tenant/TenantList.vue'
 import FloatingActionButton from '@/components/ui/FloatingActionButton.vue'
 import { useNotification } from '@/composables'
@@ -94,7 +94,7 @@ const handleTenantSubmit = async ({
       <TenantList :tenants="tenants" @select="handleCardClick" />
     </v-container>
 
-    <CreateTenantDialog
+    <TenantCreateDialog
       v-model="dialogVisible"
       :is-duplicate-name="isDuplicateName"
       @clear-duplicate-error="isDuplicateName = false"
